Add ~cuda_host_init_check to Kokkos dependency
When Legion depends on CUDA, the compile flag
"-DKOKKOS_IMPL_TURN_OFF_CUDA_HOST_INIT_CHECK" must be set. This patch
assumes that a "cuda_host_init_check" variant exists for the Kokkos
package, which is not the case currently. However, this commit has been
tested using a suitably modified version of the Kokkos Spack package.

For the record, the following modifications to the Kokkos Spack
package would be sufficient to support what is required by this commit.

74a75,77
>     variant('cuda_host_init_check', default=True,
>             description='Check host space initialization during CUDA space initialization (Disable for Legion interop)')
>
217a221,223
>
>         if "~cuda_host_init_check":
>             options.append("-DCMAKE_CXX_FLAGS=-DKOKKOS_IMPL_TURN_OFF_CUDA_HOST_INIT_CHECK")